test : 반복 일정 테스트 코드 추가
