+++ src/__mocks__/response/realEvents.json
@@ -1,1 +1,1 @@
-{"events":[{"id":"da3ca408-836a-4d98-b67a-ca389d07552b","title":"프로젝트 마감","date":"2025-08-25","startTime":"09:00","endTime":"18:00","description":"분기별 프로젝트 마감","location":"사무실","category":"업무","repeat":{"type":"none","interval":1},"notificationTime":1},{"id":"dac62941-69e5-4ec0-98cc-24c2a79a7f81","title":"생일 파티","date":"2025-08-28","startTime":"19:00","endTime":"22:00","description":"친구 생일 축하","location":"친구 집","category":"개인","repeat":{"type":"none","interval":0},"notificationTime":1},{"id":"80d85368-b4a4-47b3-b959-25171d49371f","title":"운동","date":"2025-08-22","startTime":"18:00","endTime":"19:00","description":"주간 운동","location":"헬스장","category":"개인","repeat":{"type":"none","interval":0},"notificationTime":1},{"id":"323101d3-5438-4700-80f4-e9877eac5882","title":"반복 테스트","startTime":"11:56","endTime":"12:56","description":"수정 단일 일정 변경","location":"회의실","category":"업무","notificationTime":10,"date":"2025-09-01","repeat":{"type":"daily","interval":1,"endDate":"2025-09-05","id":"694ad483-de5c-45d3-b481-73593cadb679"}},{"id":"731023ba-3068-4656-b189-c6fe93aa94b1","title":"반복 테스트","startTime":"11:56","endTime":"12:56","description":"수정 단일 일정 변경","location":"회의실","category":"업무","notificationTime":10,"date":"2025-09-02","repeat":{"type":"none","interval":1,"endDate":"2025-09-05"}},{"id":"9aab24c1-0764-40b0-a82d-f7265a3f1d09","title":"반복 테스트","startTime":"11:56","endTime":"12:56","description":"수정 단일 일정 변경","location":"회의실","category":"업무","notificationTime":10,"date":"2025-09-03","repeat":{"type":"daily","interval":1,"endDate":"2025-09-05","id":"694ad483-de5c-45d3-b481-73593cadb679"}},{"id":"1ad2becd-cb45-4d80-8956-f4099ab7544f","title":"반복 테스트","startTime":"11:56","endTime":"12:56","description":"수정 단일 일정 변경","location":"회의실","category":"업무","notificationTime":10,"date":"2025-09-04","repeat":{"type":"daily","interval":1,"endDate":"2025-09-05","id":"694ad483-de5c-45d3-b481-73593cadb679"}},{"id":"15bd4e08-10bd-4ab9-9903-284556f3e069","title":"반복 테스트","startTime":"11:56","endTime":"12:56","description":"수정 단일 일정 변경","location":"회의실","category":"업무","notificationTime":10,"date":"2025-09-05","repeat":{"type":"daily","interval":1,"endDate":"2025-09-05","id":"694ad483-de5c-45d3-b481-73593cadb679"}}]}
+{"events":[{"id":"2b7545a6-ebee-426c-b906-2329bc8d62bd","title":"팀 회의","date":"2025-08-20","startTime":"10:00","endTime":"11:00","description":"주간 팀 미팅","location":"회의실 A","category":"업무","repeat":{"type":"none","interval":0},"notificationTime":1},{"id":"09702fb3-a478-40b3-905e-9ab3c8849dcd","title":"점심 약속","date":"2025-08-21","startTime":"12:30","endTime":"13:30","description":"동료와 점심 식사","location":"회사 근처 식당","category":"개인","repeat":{"type":"none","interval":0},"notificationTime":1},{"id":"da3ca408-836a-4d98-b67a-ca389d07552b","title":"프로젝트 마감","date":"2025-08-25","startTime":"09:00","endTime":"18:00","description":"분기별 프로젝트 마감","location":"사무실","category":"업무","repeat":{"type":"none","interval":0},"notificationTime":1},{"id":"dac62941-69e5-4ec0-98cc-24c2a79a7f81","title":"생일 파티","date":"2025-08-28","startTime":"19:00","endTime":"22:00","description":"친구 생일 축하","location":"친구 집","category":"개인","repeat":{"type":"none","interval":0},"notificationTime":1},{"id":"80d85368-b4a4-47b3-b959-25171d49371f","title":"운동","date":"2025-08-22","startTime":"18:00","endTime":"19:00","description":"주간 운동","location":"헬스장","category":"개인","repeat":{"type":"none","interval":0},"notificationTime":1}]}
