+++ src/__tests__/medium.integration.spec.tsx
@@ -341,140 +341,3 @@ it('notificationTime을 10으로 하면 지정 시간 10분 전 알람 텍스트
 
   expect(screen.getByText('10분 후 기존 회의 일정이 시작됩니다.')).toBeInTheDocument();
 });
-
-it('반복 일정을 생성하면, 캘린더 뷰에 반복 아이콘이 표시된다.', async () => {
-  const events: Event[] = [
-    {
-      id: '1',
-      title: '기존 회의',
-      date: '2025-10-01',
-      startTime: '09:00',
-      endTime: '10:00',
-      description: '기존 팀 미팅',
-      location: '회의실 B',
-      category: '업무',
-      repeat: { type: 'weekly', interval: 1, endDate: '2025-10-30' },
-      notificationTime: 10,
-    },
-  ];
-  setupMockHandlerCreation(events);
-
-  setup(<App />);
-
-  const monthView = await screen.findByTestId('month-view');
-  const repeatIcons = within(monthView).getAllByTestId('repeat-icon');
-  expect(repeatIcons.length).toBe(5); // 10월 1, 8, 15, 22, 29일 등 5주
-
-  for (let day of [1, 8, 15, 22, 29]) {
-    const cell = within(monthView).getByText(String(day)).closest('td')!;
-    expect(within(cell).getByText('기존 회의')).toBeInTheDocument();
-    expect(within(cell).getByTestId('repeat-icon')).toBeInTheDocument();
-  }
-});
-
-it('반복 일정 중 하나를 수정하면 해당 일정이 단일 일정이 되고 반복 아이콘이 사라진다', async () => {
-  const mockEvents: Event[] = [
-    {
-      id: '1',
-      title: '반복 일정',
-      date: '2025-10-01',
-      startTime: '10:00',
-      endTime: '11:00',
-      description: '반복 일정 설명',
-      location: '장소',
-      category: '업무',
-      repeat: { type: 'daily', interval: 1, endDate: '2025-10-05' },
-      notificationTime: 10,
-    },
-    {
-      id: '2',
-      title: '반복 일정',
-      date: '2025-10-02',
-      startTime: '10:00',
-      endTime: '11:00',
-      description: '반복 일정 설명',
-      location: '장소',
-      category: '업무',
-      repeat: { type: 'daily', interval: 1, endDate: '2025-10-05' },
-      notificationTime: 10,
-    },
-    {
-      id: '3',
-      title: '반복 일정',
-      date: '2025-10-03',
-      startTime: '10:00',
-      endTime: '11:00',
-      description: '반복 일정 설명',
-      location: '장소',
-      category: '업무',
-      repeat: { type: 'daily', interval: 1, endDate: '2025-10-05' },
-      notificationTime: 10,
-    },
-    {
-      id: '4',
-      title: '반복 일정',
-      date: '2025-10-04',
-      startTime: '10:00',
-      endTime: '11:00',
-      description: '반복 일정 설명',
-      location: '장소',
-      category: '업무',
-      repeat: { type: 'daily', interval: 1, endDate: '2025-10-05' },
-      notificationTime: 10,
-    },
-    {
-      id: '5',
-      title: '반복 일정',
-      date: '2025-10-05',
-      startTime: '10:00',
-      endTime: '11:00',
-      description: '반복 일정 설명',
-      location: '장소',
-      category: '업무',
-      repeat: { type: 'daily', interval: 1, endDate: '2025-10-05' },
-      notificationTime: 10,
-    },
-  ];
-
-  server.use(
-    http.get('/api/events', () => {
-      return HttpResponse.json({ events: mockEvents });
-    }),
-    http.post('/api/events', async ({ request }) => {
-      const newEvent = (await request.json()) as Event;
-      newEvent.id = String(mockEvents.length + 1);
-      mockEvents.push(newEvent);
-      return HttpResponse.json(newEvent, { status: 201 });
-    }),
-    http.put('/api/events/:id', async ({ params, request }) => {
-      const { id } = params;
-      const updatedEvent = (await request.json()) as Event;
-      const index = mockEvents.findIndex((event) => event.id === id);
-
-      mockEvents[index] = { ...mockEvents[index], ...updatedEvent };
-      return HttpResponse.json(mockEvents[index]);
-    })
-  );
-  const { user } = setup(<App />);
-
-  const monthView = await screen.findByTestId('month-view');
-  let repeatIcons = within(monthView).getAllByTestId('repeat-icon');
-  expect(repeatIcons.length).toBe(5);
-
-  const editButtons = await screen.findAllByRole('button', { name: 'Edit event' });
-  await user.click(editButtons[0]); // 0번째 수정
-
-  await user.click(screen.getByLabelText('반복 일정'));
-
-  await user.click(screen.getByTestId('event-submit-button'));
-
-  repeatIcons = within(monthView).getAllByTestId('repeat-icon');
-
-  const cell = within(monthView).getByText('1').closest('td')!;
-  expect(within(cell).queryByTestId('repeat-icon')).toBeNull();
-
-  for (let day of [2, 3, 4, 5]) {
-    const repeatCell = within(monthView).getByText(String(day)).closest('td')!;
-    expect(within(repeatCell).getByTestId('repeat-icon')).toBeInTheDocument();
-  }
-});
